refactor(config): update cache to be a dedicated file, use a Defaults, create a type enum, and change maybeConfigureCluster to only override the db cache options if the setting is on redis
